OneView/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.4
google-ads==22.1.0